Ajout du message ReadChunk et son traitement par le data DataServer You can send a message ReadChunk with the fileID and the chunkID. The dataserver reply with the associated chunk (or error message)

+++ FileMetadata.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class FileMetadata implements Serializable {
     /* Taille maximale d'un chunk (en octet)
      */
-    public static final long ChunkSize = 1024 * 1024 * 4; //4MiB
+    public static final int ChunkSize = 1024 * 1024 * 4; //4MiB
     private String pathName;
     public final long id;
     private long size;
